perf: move src/types to root
