+++ src/hooks/setting/useHeaderSetting.ts
@@ -1,4 +1,4 @@
-import type { HeaderSetting } from '/@/types/config';
+import type { HeaderSetting } from '/#/config';
 
 import { computed, unref } from 'vue';
