+++ src/utils/event/index.ts
@@ -33,3 +33,10 @@ export function removeResizeListener(element: any, fn: () => any) {
     element.__ro__.disconnect();
   }
 }
+
+export function triggerWindowResize() {
+  const event = document.createEvent('HTMLEvents');
+  event.initEvent('resize', true, true);
+  (event as any).eventType = 'message';
+  window.dispatchEvent(event);
+}
