--- src/utils/event/triggerWindowResizeEvent.ts
@@ -1,9 +1,0 @@
-/**
- * triggter window.resize
- */
-export function triggerWindowResize() {
-  const event = document.createEvent('HTMLEvents');
-  event.initEvent('resize', true, true);
-  (event as any).eventType = 'message';
-  window.dispatchEvent(event);
-}
